@@ -1,5 +1,5 @@
 module github.com/DND-IT/go-imagequant
 
-go 1.20
+go 1.21
 
-require golang.org/x/image v0.6.0
+require golang.org/x/image v0.15.0
